requirements.txt
scikit-learn
xgboost
nltk
pandas
numpy
dvc
PyYAML
seaborn
matplotlib
-e .